move _validate_foreign_keys in _utils
